Move ServerConnectionSharedPreferencesPersistence to service package and rename.
