+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/bot/BotActivity.java
@@ -27,10 +27,10 @@ import com.novoda.tpbot.controls.CommandRepeater;
 import com.novoda.tpbot.controls.ControllerListener;
 import com.novoda.tpbot.controls.ControllerView;
 import com.novoda.tpbot.controls.ServerDeclarationView;
-import com.novoda.tpbot.feature_selection.ConnectedUsbDevicesFeatureSelectionController;
+import com.novoda.tpbot.bot.device.usb.ConnectedUsbDevicesFeatureSelectionController;
 import com.novoda.tpbot.feature_selection.FeatureSelectionController;
 import com.novoda.tpbot.feature_selection.FeatureSelectionPersistence;
-import com.novoda.tpbot.feature_selection.ServerConnectionSharedPreferencesPersistence;
+import com.novoda.tpbot.bot.service.ServiceConnectionSharedPreferencesPersistence;
 import com.novoda.tpbot.bot.video.calling.VideoCallSharedPreferencesPersistence;
 
 public class BotActivity extends AppCompatActivity implements BotView, DeviceConnection.DeviceConnectionListener {
@@ -53,7 +53,7 @@ public class BotActivity extends AppCompatActivity implements BotView, DeviceCon
         super.onCreate(savedInstanceState);
 
         videoCallFeature = VideoCallSharedPreferencesPersistence.newInstance(this);
-        FeatureSelectionPersistence serverConnectionFeature = ServerConnectionSharedPreferencesPersistence.newInstance(this);
+        FeatureSelectionPersistence serverConnectionFeature = ServiceConnectionSharedPreferencesPersistence.newInstance(this);
 
         setContentView(R.layout.activity_bot);
