+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/bot/service/ServiceConnectionSharedPreferencesPersistence.java
@@ -1,9 +1,11 @@
-package com.novoda.tpbot.feature_selection;
+package com.novoda.tpbot.bot.service;
 
 import android.content.Context;
 import android.content.SharedPreferences;
 
-public final class ServerConnectionSharedPreferencesPersistence implements FeatureSelectionPersistence {
+import com.novoda.tpbot.feature_selection.FeatureSelectionPersistence;
+
+public final class ServiceConnectionSharedPreferencesPersistence implements FeatureSelectionPersistence {
 
     private static final String SERVER_CONNECTION_PREF_NAME = "server_connection";
     private static final String SERVER_CONNECTION_PREFERENCES_ON_OFF_KEY = "server_connection_preferences_on_off";
@@ -12,12 +14,12 @@ public final class ServerConnectionSharedPreferencesPersistence implements Featu
 
     private final SharedPreferences sharedPreferences;
 
-    public static ServerConnectionSharedPreferencesPersistence newInstance(Context context) {
+    public static ServiceConnectionSharedPreferencesPersistence newInstance(Context context) {
         SharedPreferences sharedPreferences = context.getSharedPreferences(SERVER_CONNECTION_PREF_NAME, Context.MODE_PRIVATE);
-        return new ServerConnectionSharedPreferencesPersistence(sharedPreferences);
+        return new ServiceConnectionSharedPreferencesPersistence(sharedPreferences);
     }
 
-    private ServerConnectionSharedPreferencesPersistence(SharedPreferences sharedPreferences) {
+    private ServiceConnectionSharedPreferencesPersistence(SharedPreferences sharedPreferences) {
         this.sharedPreferences = sharedPreferences;
     }
